requirements.txt
numpy>=1.13.1
pyvcf>=0.6.8
pyfasta>=0.5.2
pysam>=0.11.2.2
